refactor: split WAV support into airspy_tools::wav
Codacy Lizard flagged rx.rs at 503 NLOC (limit 500). The WAV layer
(WavParams, header placeholder/finalize, size constants, automatic
filename) moves to its own module with its tests; every tools source
file is now well under the limit.
